replace uint by size_t in almost all places, except io function, that require sizeof(uint)

+++ include/dolfin/common/Distributed.h
@@ -16,17 +16,19 @@ class Distributed
 {
 public:
   //----------------------------------------------------------------------------
+
   Distributed( MPI::Communicator & comm );
 
   Distributed( Distributed const & other );
 
   //----------------------------------------------------------------------------
+
   // access data
-  auto       comm() -> MPI::Communicator &;
+  auto comm() -> MPI::Communicator &;
   auto comm() const -> MPI::Communicator const &;
 
-  inline auto comm_rank() const -> uint;
-  inline auto comm_size() const -> uint;
+  inline auto comm_rank() const -> size_t;
+  inline auto comm_size() const -> size_t;
 
   inline auto distributed() const -> bool;
 
@@ -46,11 +48,12 @@ protected:
 private:
   //----------------------------------------------------------------------------
   MPI::Communicator comm_;
-  uint              comm_rank_;
-  uint              comm_size_;
+  size_t            comm_rank_;
+  size_t            comm_size_;
 };
 
 //------------------------------------------------------------------------------
+
 template < typename T >
 Distributed< T >::Distributed( MPI::Communicator & comm )
   : comm_( DOLFIN_COMM_NULL )
@@ -68,11 +71,11 @@ Distributed< T >::Distributed( MPI::Communicator & comm )
 
     int ret = 0;
     MPI::check_error( MPI_Comm_rank( comm_, &ret ) );
-    comm_rank_ = static_cast< uint >( ret );
+    comm_rank_ = static_cast< size_t >( ret );
 
     ret = 0;
     MPI::check_error( MPI_Comm_size( comm_, &ret ) );
-    comm_size_ = static_cast< uint >( ret );
+    comm_size_ = static_cast< size_t >( ret );
   }
 #else
   MAYBE_UNUSED( comm );
@@ -80,6 +83,7 @@ Distributed< T >::Distributed( MPI::Communicator & comm )
 }
 
 //------------------------------------------------------------------------------
+
 template < typename T >
 Distributed< T >::Distributed( Distributed const & other )
   : comm_( DOLFIN_COMM_NULL )
@@ -89,12 +93,14 @@ Distributed< T >::Distributed( Distributed const & other )
   *this = other;
 }
 //------------------------------------------------------------------------------
+
 template < typename T >
 auto Distributed< T >::comm() -> MPI::Communicator &
 {
   return comm_;
 }
 //------------------------------------------------------------------------------
+
 template < typename T >
 auto Distributed< T >::comm() const -> MPI::Communicator const &
 {
@@ -102,20 +108,23 @@ auto Distributed< T >::comm() const -> MPI::Communicator const &
 }
 
 //------------------------------------------------------------------------------
+
 template < typename T >
-inline auto Distributed< T >::comm_rank() const -> uint
+inline auto Distributed< T >::comm_rank() const -> size_t
 {
   return comm_rank_;
 }
 
 //------------------------------------------------------------------------------
+
 template < typename T >
-inline auto Distributed< T >::comm_size() const -> uint
+inline auto Distributed< T >::comm_size() const -> size_t
 {
   return comm_size_;
 }
 
 //------------------------------------------------------------------------------
+
 template < typename T >
 inline auto Distributed< T >::distributed() const -> bool
 {
@@ -123,6 +132,7 @@ inline auto Distributed< T >::distributed() const -> bool
 }
 
 //------------------------------------------------------------------------------
+
 template < typename T >
 Distributed< T >::~Distributed()
 {
@@ -133,8 +143,10 @@ Distributed< T >::~Distributed()
 }
 
 //------------------------------------------------------------------------------
+
 template < typename T >
-auto Distributed< T >::operator=( Distributed< T > const & other ) -> Distributed< T > &
+auto Distributed< T >::operator=( Distributed< T > const & other )
+  -> Distributed< T > &
 {
   if ( this != &other )
   {
@@ -147,16 +159,18 @@ auto Distributed< T >::operator=( Distributed< T > const & other ) -> Distribute
 
       int ret = 0;
       MPI::check_error( MPI_Comm_rank( comm_, &ret ) );
-      comm_rank_ = static_cast< uint >( ret );
+      comm_rank_ = static_cast< size_t >( ret );
 
       ret = 0;
       MPI::check_error( MPI_Comm_size( comm_, &ret ) );
-      comm_size_ = static_cast< uint >( ret );
+      comm_size_ = static_cast< size_t >( ret );
     }
 #endif
   }
   return *this;
 }
+
+//------------------------------------------------------------------------------
 
 } /* namespace dolfin */
