+++ include/dolfin/la/JANPACKMat.h
@@ -23,6 +23,8 @@ namespace dolfin
 {
   class GenericSparsityPattern;
 
+//-----------------------------------------------------------------------------
+
   class JANPACKMat: public GenericMatrix, public Variable
   {
   public:
@@ -31,7 +33,7 @@ namespace dolfin
     JANPACKMat();
 
     /// Create M x N matrix
-    JANPACKMat(uint M, uint N);
+    JANPACKMat(size_t M, size_t N);
 
     /// Copy constuctor
     explicit JANPACKMat(const JANPACKMat& A);
@@ -48,7 +50,7 @@ namespace dolfin
     JANPACKMat* copy() const;
 
     /// Return size of given dimension
-    uint size(uint dim) const;
+    size_t size(size_t dim) const;
 
     /// Set all entries to zero and keep any sparse structure
     void zero();
@@ -57,39 +59,39 @@ namespace dolfin
     void apply(FinalizeType finaltype=FINALIZE);
 
     /// Display tensor
-    void disp(uint precision=2) const;
+    void disp(size_t precision=2) const;
 
     //--- Implementation of the GenericMatrix interface ---
 
     /// Initialize M x N matrix and distribute by default
-    void init(uint M, uint N);
+    void init(size_t M, size_t N);
 
     /// Initialize M x N matrix and distribute if specified
-    void init(uint M, uint N, bool distributed);
+    void init(size_t M, size_t N, bool distributed);
 
     /// Get block of values
-    void get(real* block, uint m, const uint* rows, uint n, const uint* cols) const;
+    void get(real* block, size_t m, const size_t* rows, size_t n, const size_t* cols) const;
 
     /// Set block of values
-    void set(const real* block, uint m, const uint* rows, uint n, const uint* cols);
+    void set(const real* block, size_t m, const size_t* rows, size_t n, const size_t* cols);
 
     /// Add block of values
-    void add(const real* block, uint m, const uint* rows, uint n, const uint* cols);
+    void add(const real* block, size_t m, const size_t* rows, size_t n, const size_t* cols);
 
     /// Return norm of matrix
     real norm(std::string norm_type = "frobenius") const;
 
     /// Get non-zero values of given row
-    void getrow(uint row, std::vector<uint>& columns, std::vector<real>& values) const;
+    void getrow(size_t row, std::vector<size_t>& columns, std::vector<real>& values) const;
 
     /// Set values for given row
-    void setrow(uint row, const std::vector<uint>& columns, const std::vector<real>& values);
+    void setrow(size_t row, const std::vector<size_t>& columns, const std::vector<real>& values);
 
     /// Set given rows to zero
-    void zero(uint m, const uint* rows);
+    void zero(size_t m, const size_t* rows);
 
     /// Set given rows to identity matrix
-    void ident(uint m, const uint* rows);
+    void ident(size_t m, const size_t* rows);
 
     // Matrix-vector product, y = Ax
     void mult(const GenericVector& x, GenericVector& y, bool transposed=false) const;
@@ -104,7 +106,7 @@ namespace dolfin
     const GenericMatrix& operator= (const GenericMatrix& x);
 
     /// Get number of non-zeros in the matrix
-    uint nz() const;
+    size_t nz() const;
     //--- Special functions ---
 
     /// Return linear algebra backend factory
@@ -134,8 +136,9 @@ namespace dolfin
 
   };
 
+//-----------------------------------------------------------------------------
 
-}
+} // namespace dolfin
 
 #endif
